+++ persistence-hbase/src/test/java/org/opendaylight/tsdr/datastorage/persistence/hbase/HBaseDataStoreTest.java
@@ -226,7 +226,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         TSDRMetricRecordBuilder builder2 = new TSDRMetricRecordBuilder();
 
@@ -235,7 +235,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         boolean result = true;
         try{
@@ -269,7 +269,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWTABLESTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         TSDRMetricRecordBuilder builder2 = new TSDRMetricRecordBuilder();
 
@@ -278,7 +278,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWTABLESTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         TSDRMetricRecordBuilder builder3 = new TSDRMetricRecordBuilder();
 
@@ -287,7 +287,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWTABLESTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         boolean result = true;
         try{
@@ -321,7 +321,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder2 = new TSDRMetricRecordBuilder();
@@ -331,7 +331,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder3 = new TSDRMetricRecordBuilder();
@@ -341,7 +341,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder4 = new TSDRMetricRecordBuilder();
@@ -351,7 +351,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder5 = new TSDRMetricRecordBuilder();
@@ -361,7 +361,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder6 = new TSDRMetricRecordBuilder();
@@ -371,7 +371,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder7 = new TSDRMetricRecordBuilder();
@@ -381,7 +381,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder8 = new TSDRMetricRecordBuilder();
@@ -391,7 +391,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder9 = new TSDRMetricRecordBuilder();
@@ -401,7 +401,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         TSDRMetricRecordBuilder builder10 = new TSDRMetricRecordBuilder();
 
@@ -410,7 +410,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder11 = new TSDRMetricRecordBuilder();
@@ -420,7 +420,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder12 = new TSDRMetricRecordBuilder();
@@ -430,7 +430,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         TSDRMetricRecordBuilder builder13 = new TSDRMetricRecordBuilder();
 
@@ -439,7 +439,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
 
         TSDRMetricRecordBuilder builder14 = new TSDRMetricRecordBuilder();
@@ -449,7 +449,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.PORTSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         boolean result = true; 
         try{
@@ -497,7 +497,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.QUEUESTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
        TSDRMetricRecordBuilder builder2 = new TSDRMetricRecordBuilder();
 
@@ -506,7 +506,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.QUEUESTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
        TSDRMetricRecordBuilder builder3 = new TSDRMetricRecordBuilder();
 
@@ -515,7 +515,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.QUEUESTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         boolean result = true;
 
@@ -555,7 +555,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWMETERSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
        TSDRMetricRecordBuilder builder2 = new TSDRMetricRecordBuilder();
 
@@ -564,7 +564,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWMETERSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
        TSDRMetricRecordBuilder builder3 = new TSDRMetricRecordBuilder();
 
@@ -573,7 +573,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWMETERSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         boolean result = true;
         try{
@@ -612,7 +612,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWGROUPSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
        TSDRMetricRecordBuilder builder2 = new TSDRMetricRecordBuilder();
 
@@ -621,7 +621,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWGROUPSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
        TSDRMetricRecordBuilder builder3 = new TSDRMetricRecordBuilder();
 
@@ -630,7 +630,7 @@ public class HBaseDataStoreTest {
             .setNodeID("node1")
             .setRecordKeys(recordKeys)
             .setTSDRDataCategory(DataCategory.FLOWGROUPSTATS)
-            .setTimeStamp(new BigInteger(timeStamp)).build();
+            .setTimeStamp(new Long(timeStamp)).build();
 
         boolean result = true;
